+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## To Be Released
 
+* feat(apps): support HDS resource on application creation
+
 ## 1.42.0
 
 * feat(databases next generation): add firewall rules management commands: